requirements.txt
asgiref==3.7.2
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
numpy
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
PyJWT==2.8.0
pytesseract==0.3.10
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
gunicorn==21.2.0
django-cors-headers==4.2.0